Bug fixes and fixed compiler warnings with -Wconversion
pthread_create() call was assigning the status to the thread heandle, causing
pthread_detach() to fail. The rest of the changes were assignment warnigns
found with -Wconversion.

Added placeholder for VDIR_EVENT_RESPONSE type used by
WatchSendEvents().

+++ lwraft/server/grpc-head/postwatch/postwatch_server_c.c
@@ -14,7 +14,7 @@
 #include "includes.h"
 
 typedef struct _VMDIR_GRPC_THREAD_CONTEXT {
-    DWORD argc;
+    int argc;
     PSTR *argv;
 } VMDIR_GRPC_THREAD_CONTEXT, *PVMDIR_GRPC_THREAD_CONTEXT;
 
@@ -39,7 +39,7 @@ DWORD
 VmDirGrpcPostWatchServerInitialize(void)
 {
     DWORD dwError = 0;
-    DWORD argc = 2;
+    int argc = 2;
     VMDIR_THREAD thread = {0};
     PVMDIR_GRPC_THREAD_CONTEXT pGrpcContext = NULL;
     PSTR *ppArgv = NULL;
@@ -55,7 +55,7 @@ VmDirGrpcPostWatchServerInitialize(void)
     BAIL_ON_VMDIR_ERROR(dwError);
 
     dwError = VmDirAllocateMemory(
-                sizeof(PSTR) * argc,
+                sizeof(PSTR) * (size_t) argc,
                 (PVOID*)&ppArgv);
     BAIL_ON_VMDIR_ERROR(dwError);
